+++ src/apollo/index.ts
@@ -1,4 +1,5 @@
 export { default as Navigation } from "./navigation/Navigation";
+export { default as Notification } from "./notification/Notification";
 export { default as PageContent } from "./pageContent/PageContent";
 export { PageDocument } from "../common/headlessService/page";
 export { MenuDocument } from "../common/headlessService/menu";